pyproject.toml
# SPDX-License-Identifier: MIT
# Copyright (c) 2025 Advanced Micro Devices, Inc. All rights reserved.

[build-system]
requires = ["setuptools>=61", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "iris"
dynamic = ["version"]
description = "Triton-based framework for Remote Memory Access (RMA) operations with SHMEM-like APIs for multi-GPU programming."
authors = [
    { name = "Muhammad Awad", email = "muhaawad@amd.com" },
    { name = "Muhammad Osama", email = "Muhammad.Osama@amd.com" },
    { name = "Brandon Potter", email = "Brandon.Potter@amd.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "requests",
    "ruff",
    "tritonblas @ git+https://github.com/ROCm/tritonBLAS.git@df58476a4520b72495a3f03f911368a184126568",

]

[project.urls]
Homepage = "https://rocm.github.io/iris/"
Repository = "https://github.com/ROCm/iris"
Documentation = "https://rocm.github.io/iris/"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "mypy",
]

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
include = ["iris*"]

# ---- setuptools-scm versioning ----
[tool.setuptools_scm]
version_scheme = "post-release"   # .postN after last tag
local_scheme = "node-and-date"    # add commit hash (e.g. +gabc1234) and date (e.g. +20250914)
fallback_version = "0.0.0"        # used if git metadata unavailable

[tool.ruff]
line-length = 120
exclude = [
    "**/*.ipynb"                  # match notebooks anywhere
]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501", "E701", "E731", "E741", "F841"]

[tool.ruff.format]
quote-style = "double"